Fix lib dependencies in web to run app
Dockerfile base on node:alpine and need to install dependencies in guanyu-core after install dependencies in web.
web dependencies add guanyu-core by file, and nodemon in devDependencies.
add web interface v1 and v2
