+++ web/Dockerfile
@@ -1,10 +1,11 @@
-FROM clifflu/sophos-av-npm
+FROM node:alpine
 
-COPY ./guanyu /guanyu
-WORKDIR /guanyu
+COPY web /web
+COPY guanyu-core /guanyu-core
 
-RUN npm install --production && npm dedupe
-RUN sed -i "s/<Email><Status>enabled/<Email><Status>disabled/" /opt/sophos-av/etc/savd.cfg
+RUN cd /web && npm install --production && npm dedupe
+RUN cd /guanyu-core && npm install --production && npm dedupe
+WORKDIR /web
 
 EXPOSE 3000
