+++ web/Dockerfile.dev
@@ -1,0 +1,12 @@
+FROM node:alpine
+
+COPY web /web
+COPY guanyu-core /guanyu-core
+
+RUN cd /web && npm install && npm dedupe
+RUN cd /guanyu-core && npm install && npm dedupe
+WORKDIR /web
+
+EXPOSE 3000
+
+CMD ["npm", "run", "dev"]
