+++ web/package.json
@@ -3,9 +3,9 @@
   "version": "0.2.0-dev",
   "private": true,
   "scripts": {
-    "dev": "nodemon bin/www",
+    "dev": "nodemon app.js",
     "prestart": "/opt/sophos-av/bin/savdctl start || echo '** Running without Sophos-AV'",
-    "start": "node bin/www",
+    "start": "node app.js",
     "cov": "node_modules/.bin/istanbul cover --dir ./build/coverage node_modules/.bin/_mocha -- -R spec --recursive && [ -x `which open` ] && open build/coverage/lcov-report/index.html"
   },
   "dependencies": {
@@ -24,9 +24,11 @@
     "semaphore": "^1.1.0",
     "serve-favicon": "^2.3.2",
     "tmp": "0.0.28",
-    "winston": "^2.3.1"
+    "winston": "^2.3.1",
+    "guanyu-core": "file:../guanyu-core"
   },
   "devDependencies": {
+    "nodemon": "^1.18.4",
     "chai": "^3.5.0",
     "istanbul": "^0.4.5",
     "mocha": "^2.5.3",
