+++ web/views/usage-scan.jade
@@ -3,4 +3,8 @@ extends layout
 block content
     h1 Scan what?
 
-    p #[a(href="file") file], #[a(href="text") text], or #[a(href="uri") uri]?
+    h2 Version 1:
+    p #[a(href="v1/file") file], #[a(href="v1/text") text], or #[a(href="v1/uri") uri]?
+
+    h2 Version 2:
+    p #[a(href="v2/file") file], #[a(href="v2/text") text], or #[a(href="v2/uri") uri]?
